api/exceptions.py
from django.utils.translation import gettext_lazy as _

from rest_framework import status
from rest_framework.exceptions import APIException


class ServiceError(APIException):
    status_code = status.HTTP_503_SERVICE_UNAVAILABLE
    default_detail = _('Service Unavailable.')
    default_code = 'service_unavailable'


class ThirdPartyError(APIException):
    status_code = status.HTTP_503_SERVICE_UNAVAILABLE
    default_detail = _('Service Unavailable.')
    default_code = 'service_unavailable'
