requirements.txt
asgiref==3.5.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
idna==3.3
inflection==0.5.1
jsonschema==4.7.2
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.10